Finished Rev3
Redesigned the scatter from rev2 to be the implementation that he gave in class.  This allowed the MPI_Scatterv to work properly with the rest of the code.  All three versions seem to work, need to make final edits to the code and batch them on the cluster.  Also, finish the report and turn in the assignment.