Safe Commit
This current version works with smaller data sets, but the copying of the columns in the B matrix get to large when the data set is large.  This is a safe commit before changing the code to broadcast the entire B matrix to every process, to save on space and copy steps.