Broadcast branch
The main now broadcasts be to all, and there are no more heap errors.  But the indexing into b for the calculation still seems to be wrong as there are a few very large numbers that look like garbage, or something outside of the memory space of the matrix.  I need to fix that and everything should be good to go.