Final Code Edits and Started Report
Made the final edits to the code and fixed a bug on the column read where it was using dim_m instead of dim_n.  This caused it to work in cases where dim_m = dim_n, but failed one of my other test cases.  Finish the report and HW should be done.