Updated MPI Implementation and Finished OpenMP Implemenation
Updated the OpenMPI implemenation but there is some optimization that
needs to be done. I need to be able to send a done flag to all processes
so that they can exit their loop.  This way I can finish without
processing the rest of the dictionary file.  The OpenMP version seems to
be finished and working, further testing my show bugs but it seems to be
doing very well right now.  Need to run a bunch of tests with these and
get some charts.  Need to start the report.